refactor(data-store): extract sqlite-schema from sqlite-adapter
Moves the table-creation DDL and version-migration runner statements out of sqlite-adapter into sqlite-schema, behavior unchanged.
One atomic slice of the step-4 large-file decomposition.
